Decomplixify logic by isolating data-sorting utility function
Previously, the logic used to sort our regions by a give key (eg,
percentage of population vaccinated), AS WELL AS the calculation
for the percentages was all bundled up in one function.

What that meant was that we were passing in 2 data structures in a
few places:

1. With the actual data we wanted (eg, province name, population, etc)
2. With the sort order and percentage number

Changed up the logic so that percentages are included in the data
about the regions, all good.

Also pulled out the "sort" function so that it can be used for
other data structures.

It's hard to explain how cool this is, but it's awesome.