Update "methodology" and "about" pages with latest info
The latest news about this is that we are hoping for ... 15 million
by Canada Day (jeez).

That amounts to ~38%, which would be a pretty big jump considering
how slow it's all been going.

The 70% by September idea wasn't really based on anything super
concrete, so let's just drop it and focus on the new news.